Break app into several routers
Breaks the app object into several routers for different concerns:
- platform: documentation, configuration, monitoring, metrics, and authn
- stripe:  ingesting stripe objects
- contacts: contact management, including newsletters and waitlists

Also adds a dedicated dependencies module for common dependencies used
by multiple routers
