+++ ctms/metrics.py
@@ -4,11 +4,13 @@ from itertools import product
 from typing import Any, Type, cast
 
 from fastapi import FastAPI
+from fastapi.security import HTTPBasic
 from prometheus_client import CollectorRegistry, Counter, Histogram
 from prometheus_client.utils import INF
 from sqlalchemy.orm import Session
 from starlette.routing import Route
 
+from ctms.auth import OAuth2ClientCredentials
 from ctms.crud import get_active_api_client_ids
 
 METRICS_PARAMS: dict[str, tuple[Type[Counter] | Type[Histogram], dict]] = {
@@ -62,6 +64,24 @@ METRICS_PARAMS: dict[str, tuple[Type[Counter] | Type[Histogram], dict]] = {
         },
     ),
 }
+
+# We could use the default prometheus_client.REGISTRY, but it makes tests
+# easier to write if it is possible to replace the registry with a fresh one.
+METRICS_REGISTRY = CollectorRegistry()
+METRICS = None
+
+
+def set_metrics(metrics: Any) -> None:
+    global METRICS  # pylint: disable=global-statement
+    METRICS = metrics
+
+
+# pylint: disable-next=unnecessary-lambda-assignment
+get_metrics_registry = lambda: METRICS_REGISTRY
+get_metrics = lambda: METRICS  # pylint: disable=unnecessary-lambda-assignment
+
+oauth2_scheme = OAuth2ClientCredentials(tokenUrl="token")
+token_scheme = HTTPBasic(auto_error=False)
 
 
 def init_metrics(registry: CollectorRegistry) -> dict[str, Counter | Histogram]:
